WAS Liberty workarounds
Calculate full URL with scheme for temporaryRedirect

WebSphere Liberty does not redirect correctly when the URI just starts
with "/" it requires a full URL with scheme in order to redirect as is
rather than making it relative to the base URI of the JAX-RS
application.

Added blank token and userinfo files.  WebSphere Liberty does not
implement SEND_CONTINUE correctly.  Putting these files in will allow
the /token and /userinfo pages to appear.